Fixes for BigDecimal that takes in int into the constructor. Does not work on jdk1.4 when compiled on jdk1.5

+++ src/com/idega/block/finance/presentation/TariffEditor.java
@@ -299,7 +299,7 @@ public class TariffEditor extends Finance {
 						Map.Entry entry = (Map.Entry) iter.next();
 						String tatt = (String) entry.getKey();
 						String val = "";
-						BigDecimal attTotal = new BigDecimal(0);
+						BigDecimal attTotal = new BigDecimal((double)0);
 
 						List tariffsInMap = (List) entry.getValue();
 						for (Iterator iterator = tariffsInMap.iterator(); iterator.hasNext();) {
